+++ Core/Engine/TimeStep.h
@@ -2,13 +2,13 @@
 class TimeStep
 {
 private:
-	double m_CurrentTime;
+	float m_CurrentTime;
 
 public:
 	TimeStep() : m_CurrentTime{0} {}
-	operator double()  { return m_CurrentTime; }
+	operator float()  { return m_CurrentTime; }
 
-	double& operator=(const double& val) {
+	float& operator=(const float& val) {
 		this->m_CurrentTime = val;
 		return this->m_CurrentTime;
 	}
